web/requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
h5py==3.12.1
importlib_resources==6.5.2
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2025.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
zipp==3.21.0
